\name{miRTargetForSpecies}

\alias{miRTargetForSpecies-package}
\alias{miRTargetForSpecies}

\title{miRTargetsForSpecies : Collection of miRNA Target Prediction Data}

\description{
  This is a data collection for computationally predicted miRNA
         targets for various species including Homo sapiens,
			   Mus musculus, Rattus norvegicus, Drosophila melanogaster,
			   Danio rerio, Caenorhabditis elegans. Other than Rattus norvegicus,
			   we collect target predictions from TargetScan and for Rattus
			   norvegicus, miRmap database predictions are utilized.
}
